Created pushSkill function
Created pushSkill function to push a skill that user entered in the skill input. Also, each time user enters a skill in the input, it creates a little card below the skill input to show the user what he has added. Will add a way to remove the skill from the CV and the little card section.